requirements.txt
opencv-python
mediapipe
comtypes
numpy
pycaw
paho-mqtt